# Conflicts:
#	Frameworks/Frameworks/x64/Debug/Framework01.tlog/CL.command.1.tlog
#	Frameworks/Frameworks/x64/Debug/Framework01.tlog/CL.read.1.tlog
#	Frameworks/Frameworks/x64/Debug/Framework01.tlog/CL.write.1.tlog
#	Frameworks/Frameworks/x64/Debug/Framework01.tlog/Framework01.lastbuildstate
#	Frameworks/Frameworks/x64/Debug/Framework01.tlog/link.read.1.tlog
#	Frameworks/Frameworks/x64/Debug/Framework01.tlog/link.write.1.tlog
#	Frameworks/Frameworks/x64/Debug/Framework01.tlog/rc.read.1.tlog
#	Frameworks/Frameworks/x64/Debug/Framework01.tlog/rc.write.1.tlog
#	Frameworks/Frameworks/x64/Debug/Frameworks.log
#	Frameworks/Frameworks/x64/Debug/vc141.idb
#	Frameworks/Frameworks/x64/Debug/vc141.pdb
#	Frameworks/x64/Debug/Framework01.ilk
#	Frameworks/x64/Debug/Framework01.pdb

+++ Frameworks/Frameworks/Code/00.Global/01.Utility/Vectors.h
@@ -38,12 +38,12 @@ inline int RandInt(int x, int y)
 	return rand() % (y - x + 1) + x;
 }
 
-inline double RandFloat() 
+inline float RandFloat() 
 { 
 	return ((rand()) / (RAND_MAX + 1.0));
 }
 
-inline double RandInRange(double x, double y)
+inline float RandInRange(float x, float y)
 {
 	return x + RandFloat()*(y - x);
 }
@@ -60,10 +60,10 @@ inline double RandomClamped()
 	return RandFloat() - RandFloat(); 
 }
 
-inline double RandGaussian(double mean = 0.0, double standard_deviation = 1.0)
+inline float RandGaussian(float mean = 0.0, float standard_deviation = 1.0)
 {
-	double x1, x2, w, y1;
-	static double y2;
+	float x1, x2, w, y1;
+	static float y2;
 	static int use_last = 0;
 
 	if (use_last)		        /* use value from previous call */
